requirements.txt
aiocontextvars==0.2.2
aiohttp==3.6.2
alembic==1.4.2
APScheduler==3.6.3
arrow==0.15.5
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
boto3==1.12.36
boto3-stubs==1.12.36.0
botocore==1.15.36
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
configparser==5.0.0
confuse==1.0.0
contextvars==2.4
coverage==5.0.4
cryptography==2.9
dataclasses==0.6
DataProperty==0.47.0
dnspython==1.15.0
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Limiter==1.3.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
immutables==0.11
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsons==1.2.0
jsonschema==3.2.0
limits==1.5.1
loguru==0.5.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mbstrdecoder==1.0.0
mimerender==0.6.0
msgfy==0.0.7
multidict==4.7.6
mypy-boto3==1.12.36.0
nassl==2.2.0
notifiers==1.2.1
path==13.2.0
path.py==12.4.0
pathvalidate==2.2.2
pshtt==0.6.6
publicsuffix==1.1.1
py3dns==3.2.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pyspf==2.0.14
pytablereader==0.29.0
pytablewriter==0.50.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
rq==1.3.0
rq-dashboard==0.6.1
s3transfer==0.3.3
six==1.14.0
slackclient==2.7.1
soupsieve==2.0
SQLAlchemy==1.3.15
sslyze==2.1.4
strict-rfc3339==0.7
supervisor==4.2.0
tabledata==1.1.1
tls-parser==1.2.1
toml==0.10.0
tqdm==4.45.0
trustymail==0.7.5
typepy==1.0.0
typing-extensions==3.7.4.2
typish==1.7.0
tzlocal==2.1
urllib3==1.25.8
Werkzeug==1.0.1
wget==3.2
win32-setctime==1.0.1
WTForms==2.2.1
yarl==1.4.2
zipp==3.1.0
